+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - (Maintenance) Deprecate --operator.ml CLI flag (hidden, warns if used)
 - (Feature) Remove Analytics operator and GraphAnalyticsEngine CRD
 - (Maintenance) Deprecate --operator.analytics CLI flag (hidden, warns if used)
+- (Feature) Remove Apps operator and ArangoJob CRD
+- (Maintenance) Deprecate --operator.apps CLI flag (hidden, warns if used)
 
 ## [1.4.3](https://github.com/arangodb/kube-arangodb/tree/1.4.3) (2026-04-23)
 - (Bugfix) Fix Community image discovery